Consolidate Triton helpers
